[modmail] extract methods to own class
